Remodulação doa componentes
